% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kg_plot_palette.R
\name{convert_colTodf}
\alias{convert_colTodf}
\title{Convert color data to list class}
\usage{
convert_colTodf(coldata)
}
\arguments{
\item{coldata}{Color data.}
}
\value{
list class containing palette name (or color name),
original color data (data.frame), column name, and row name.
}
\description{
\code{convert_colTodf} This function convert color data to list class.
}
\examples{
convert_colTodf(coldata = c("red", "#ff48ac"))
}
