% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KaradaColor_Package.R
\docType{data}
\name{df_KG_palettes}
\alias{df_KG_palettes}
\title{Color palette dataset for the KaradaColor Package}
\format{
An object of class \code{list} of length 3.
}
\usage{
df_KG_palettes
}
\value{
Vector containing a hex color code and palette name
}
\description{
\code{df_KG_palettes} Displays the color palette dataset
}
\examples{
data("df_KG_palettes")
head(df_KG_palettes)
}
\keyword{data}
