% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Use_KG_Palette.R
\name{get_KG_color}
\alias{get_KG_color}
\alias{get_KG_palette}
\title{Get color palette data set for the KaradaColor Package}
\usage{
get_KG_color(n = 5, name = "Sapporo_Sta", alpha = 1)

get_KG_palette(name = "Sapporo_Sta", alpha = 1)
}
\arguments{
\item{n}{Number of colors to create.}

\item{name}{color palette name. (default: \code{"Sapporo_Sta"})}

\item{alpha}{The alpha transparency. Numbers range from 0-1. see argument alpha in \code{\link[grDevices]{rgb}}.}
}
\value{
Vector containing a hex color code representation for the chosen palette.
}
\description{
\code{get_KG_color} Get color palette data set for the KaradaColor Package
}
\examples{
library("scales")
show_col(get_KG_color(n = 20, name = "Totan", alpha = 0.5))
library("aRtsy")
set.seed(1234)
canvas_strokes(colors = get_KG_color(name = "Otaru_Unga"),
               neighbors = 1, p = 0.02, iterations = 1,
               resolution = 350)
}
\author{
KARADA-GOOD \email{info@karada-good.net}, \href{https://github.com/KaradaGood/KaradaColor}{Github}, \href{https://www.karada-good.net/analyticsr/r-770/}{Personal Blog}
}
