% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_Color_Schemes.R
\name{KG_create_cc}
\alias{KG_create_cc}
\title{Create complementary color}
\usage{
KG_create_cc(basecols = c("#ff48ac", "red"), alpha = 1)
}
\arguments{
\item{basecols}{Set basecol color data.}

\item{alpha}{The alpha transparency. Numbers range from 0-1. see argument alpha in \code{\link[grDevices]{rgb}}.}
}
\value{
Return base and complementary color in a data frame
}
\description{
\code{show_KG_palette} This function create complementary color.
}
\examples{
show_KG_palette(KG_create_cc())
plot_KG_palette(color = KG_create_cc())
}
\author{
Masaki Fujishima \email{info@karada-good.net}, \href{https://github.com/KaradaGood/KaradaColor}{Github}, \href{https://www.karada-good.net/analyticsr/r-770/}{Personal Blog}
}
