% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kg_color_schemes.R
\name{kg_create_mc}
\alias{kg_create_mc}
\alias{range_RGB}
\title{Monochromatic Color}
\usage{
kg_create_mc(basecols = "#7F36EF", n = 2, alpha = 1)

range_RGB(basecols = NULL, n = NULL)
}
\arguments{
\item{basecols}{Color data.}

\item{n}{Create n colors of low and high saturation
with Color data in the center.}

\item{alpha}{The alpha transparency. Numbers range from 0-1.
see argument alpha in \code{\link[grDevices]{rgb}}.#'}
}
\value{
Return monochromatic color in a data frame
}
\description{
\code{kg_create_mc} This function create monochromatic color.
}
\examples{
kg_show_console(kg_create_mc())
kg_plot_color(color = kg_create_mc(basecols = "#d0af4c"))
}
