% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kg_color_schemes.R
\name{kg_create_tc}
\alias{kg_create_tc}
\title{Create triad color}
\usage{
kg_create_tc(basecols = "#4E5C78", alpha = 1)
}
\arguments{
\item{basecols}{Set basecol color data.}

\item{alpha}{The alpha transparency. Numbers range from 0-1.
see argument alpha in \code{\link[grDevices]{rgb}}.}
}
\value{
Return triad color in a data frame
}
\description{
\code{kg_create_tc} This function create triad color.
}
\examples{
kg_show_console(kg_create_tc())
kg_plot_color(color = kg_create_tc())
}
