% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kg_show_palette.R
\name{kg_show_consol}
\alias{kg_show_consol}
\alias{console_col}
\title{Display the color palette data set in the consol}
\usage{
kg_show_consol(x = df_KG_palettes)

console_col(x)
}
\arguments{
\item{x}{color palette data.}
}
\value{
Vector containing a hex color code representation for the chosen palette
}
\description{
\code{kg_show_consol} This function draws display the color palette data set
in the console.
}
\examples{
kg_show_consol()
}
