% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show_KG_Palette.R
\name{Show_KG_Palette}
\alias{Show_KG_Palette}
\title{Display the color palette data set for the KaradaColor Package in the console}
\usage{
Show_KG_Palette(x = df_KG_palettes)
}
\value{
Vector containing a hex color code representation for the chosen palette
}
\description{
\code{Show_KG_Palette} Display the color palette data set for the KaradaColor Package in the console
}
\examples{
Show_KG_Palette()
}
