% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kg_plot_palette.R
\name{image_add_textAcode}
\alias{image_add_textAcode}
\title{Add to image() command text and color code}
\usage{
image_add_textAcode(
  coldata,
  xlab = NULL,
  ylab = NULL,
  colornames = "color",
  showcode
)
}
\arguments{
\item{coldata}{Color data.}

\item{xlab}{x axis label.}

\item{ylab}{x axis label.}

\item{colornames}{Color name.}

\item{showcode}{Add a color code to the plot.}
}
\description{
\code{image_add_textAcode} This function add to image() command text and color code.
}
