% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kg_color_schemes.R
\name{kg_create_cc}
\alias{kg_create_cc}
\title{Create complementary color}
\usage{
kg_create_cc(basecols = c("#ff48ac", "red"), alpha = 1)
}
\arguments{
\item{basecols}{Set color data.}

\item{alpha}{The alpha transparency. Numbers range from 0-1.
see argument alpha in \code{\link[grDevices]{rgb}}.}
}
\value{
Return base and complementary color in a data frame
}
\description{
\code{kg_create_cc} This function create complementary color.
}
\examples{
kg_show_console(kg_create_cc())
kg_plot_color(color = kg_create_cc())
}
