% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KG_Color_Schemes.R
\name{KG_create_tc}
\alias{KG_create_tc}
\title{Create triad color}
\usage{
KG_create_tc(basecols = "#4E5C78", alpha = 1)
}
\arguments{
\item{basecols}{Set basecol color data.}

\item{alpha}{The alpha transparency. Numbers range from 0-1. see argument alpha in \code{\link[grDevices]{rgb}}.}
}
\value{
Return triad color in a data frame
}
\description{
\code{show_KG_palette} This function create triad color.
}
\examples{
show_KG_palette(KG_create_tc())
plot_KG_palette(color = KG_create_tc())
}
\author{
Masaki Fujishima \email{info@karada-good.net}, \href{https://github.com/KaradaGood/KaradaColor}{Github}, \href{https://www.karada-good.net/analyticsr/r-770/}{Personal Blog}
}
