% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_KG_Palette.R
\name{plot_KG_palette}
\alias{plot_KG_palette}
\title{Draw the color palette of the KaradaColor Package}
\usage{
plot_KG_palette(n = 5, name = "Sapporo_Sta", colplot = TRUE)
}
\arguments{
\item{n}{Number of colors to create.}

\item{name}{color palette name. (default: \code{"Sapporo_Sta"})}

\item{colplot}{Plotting Color Palette.}
}
\value{
Vector containing a hex color code representation for the chosen palette
}
\description{
\code{plot_KG_Palette} This function draw the color palette of the KaradaColor Package.
}
\examples{
plot_KG_palette(name = "Amanohashidate", n = 5, colplot = TRUE)
library("ggplot2")
ggplot(data = data.frame(x = 1:5, y = 1:5), aes(x, y)) +
  geom_point(col = plot_KG_palette("Sapporo_Sta", n = 5,
                                   colplot = FALSE),
                                   size = 10)
}
\author{
Masaki Fujishima \email{info@karada-good.net}, \href{https://github.com/KaradaGood/KaradaColor}{Github}, \href{https://www.karada-good.net/analyticsr/r-770/}{Personal Blog}
}
