% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_KG_Palette.R
\name{Plot_KG_Palette}
\alias{Plot_KG_Palette}
\title{Displays the color palette of the KaradaColor Package}
\usage{
Plot_KG_Palette(name, n = 5, colplot = TRUE)
}
\arguments{
\item{name}{color palette name}

\item{n}{Number of colors to create}

\item{colplot}{Plotting Color Palette}
}
\value{
Vector containing a hex color code representation for the chosen palette
}
\description{
\code{Plot_KG_Palette} Displays the color palette of the KaradaColor Package
}
\examples{
Plot_KG_Palette("Amanohashidate", n = 5, colplot = TRUE)
}
